reorganized urls for enterprise and orgs endpoints for action permissions
